Move other Appmesh structure functions to resource files.
