+++ aws/structure.go
@@ -13,7 +13,6 @@ import (
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/private/protocol/json/jsonutil"
 	"github.com/aws/aws-sdk-go/service/apigateway"
-	"github.com/aws/aws-sdk-go/service/appmesh"
 	"github.com/aws/aws-sdk-go/service/appsync"
 	"github.com/aws/aws-sdk-go/service/autoscaling"
 	"github.com/aws/aws-sdk-go/service/cloudformation"
@@ -4826,446 +4825,6 @@ func flattenRdsScalingConfigurationInfo(scalingConfigurationInfo *rds.ScalingCon
 	}
 
 	return []interface{}{m}
-}
-
-func expandAppmeshMeshSpec(vSpec []interface{}) *appmesh.MeshSpec {
-	spec := &appmesh.MeshSpec{}
-
-	if len(vSpec) == 0 || vSpec[0] == nil {
-		// Empty Spec is allowed.
-		return spec
-	}
-	mSpec := vSpec[0].(map[string]interface{})
-
-	if vEgressFilter, ok := mSpec["egress_filter"].([]interface{}); ok && len(vEgressFilter) > 0 && vEgressFilter[0] != nil {
-		mEgressFilter := vEgressFilter[0].(map[string]interface{})
-
-		if vType, ok := mEgressFilter["type"].(string); ok && vType != "" {
-			spec.EgressFilter = &appmesh.EgressFilter{
-				Type: aws.String(vType),
-			}
-		}
-	}
-
-	return spec
-}
-
-func flattenAppmeshMeshSpec(spec *appmesh.MeshSpec) []interface{} {
-	if spec == nil {
-		return []interface{}{}
-	}
-
-	mSpec := map[string]interface{}{}
-
-	if spec.EgressFilter != nil {
-		mSpec["egress_filter"] = []interface{}{
-			map[string]interface{}{
-				"type": aws.StringValue(spec.EgressFilter.Type),
-			},
-		}
-	}
-
-	return []interface{}{mSpec}
-}
-
-func expandAppmeshVirtualRouterSpec(vSpec []interface{}) *appmesh.VirtualRouterSpec {
-	spec := &appmesh.VirtualRouterSpec{}
-
-	if len(vSpec) == 0 || vSpec[0] == nil {
-		// Empty Spec is allowed.
-		return spec
-	}
-	mSpec := vSpec[0].(map[string]interface{})
-
-	if vListeners, ok := mSpec["listener"].(*schema.Set); ok && vListeners.Len() > 0 {
-		listeners := []*appmesh.VirtualRouterListener{}
-
-		for _, vListener := range vListeners.List() {
-			listener := &appmesh.VirtualRouterListener{}
-
-			mListener := vListener.(map[string]interface{})
-
-			if vPortMapping, ok := mListener["port_mapping"].([]interface{}); ok && len(vPortMapping) > 0 && vPortMapping[0] != nil {
-				mPortMapping := vPortMapping[0].(map[string]interface{})
-
-				listener.PortMapping = &appmesh.PortMapping{}
-
-				if vPort, ok := mPortMapping["port"].(int); ok && vPort > 0 {
-					listener.PortMapping.Port = aws.Int64(int64(vPort))
-				}
-				if vProtocol, ok := mPortMapping["protocol"].(string); ok && vProtocol != "" {
-					listener.PortMapping.Protocol = aws.String(vProtocol)
-				}
-			}
-
-			listeners = append(listeners, listener)
-		}
-
-		spec.Listeners = listeners
-	}
-
-	return spec
-}
-
-func flattenAppmeshVirtualRouterSpec(spec *appmesh.VirtualRouterSpec) []interface{} {
-	if spec == nil {
-		return []interface{}{}
-	}
-
-	mSpec := map[string]interface{}{}
-
-	if spec.Listeners != nil {
-		vListeners := []interface{}{}
-
-		for _, listener := range spec.Listeners {
-			mListener := map[string]interface{}{}
-
-			if listener.PortMapping != nil {
-				mPortMapping := map[string]interface{}{
-					"port":     int(aws.Int64Value(listener.PortMapping.Port)),
-					"protocol": aws.StringValue(listener.PortMapping.Protocol),
-				}
-				mListener["port_mapping"] = []interface{}{mPortMapping}
-			}
-
-			vListeners = append(vListeners, mListener)
-		}
-
-		mSpec["listener"] = schema.NewSet(appmeshVirtualNodeListenerHash, vListeners)
-	}
-
-	return []interface{}{mSpec}
-}
-
-func expandAppmeshVirtualNodeSpec(vSpec []interface{}) *appmesh.VirtualNodeSpec {
-	spec := &appmesh.VirtualNodeSpec{}
-
-	if len(vSpec) == 0 || vSpec[0] == nil {
-		// Empty Spec is allowed.
-		return spec
-	}
-	mSpec := vSpec[0].(map[string]interface{})
-
-	if vBackends, ok := mSpec["backend"].(*schema.Set); ok && vBackends.Len() > 0 {
-		backends := []*appmesh.Backend{}
-
-		for _, vBackend := range vBackends.List() {
-			backend := &appmesh.Backend{}
-
-			mBackend := vBackend.(map[string]interface{})
-
-			if vVirtualService, ok := mBackend["virtual_service"].([]interface{}); ok && len(vVirtualService) > 0 && vVirtualService[0] != nil {
-				mVirtualService := vVirtualService[0].(map[string]interface{})
-
-				backend.VirtualService = &appmesh.VirtualServiceBackend{}
-
-				if vVirtualServiceName, ok := mVirtualService["virtual_service_name"].(string); ok {
-					backend.VirtualService.VirtualServiceName = aws.String(vVirtualServiceName)
-				}
-			}
-			backends = append(backends, backend)
-		}
-
-		spec.Backends = backends
-	}
-
-	if vListeners, ok := mSpec["listener"].(*schema.Set); ok && vListeners.Len() > 0 {
-		listeners := []*appmesh.Listener{}
-
-		for _, vListener := range vListeners.List() {
-			listener := &appmesh.Listener{}
-
-			mListener := vListener.(map[string]interface{})
-
-			if vHealthCheck, ok := mListener["health_check"].([]interface{}); ok && len(vHealthCheck) > 0 && vHealthCheck[0] != nil {
-				mHealthCheck := vHealthCheck[0].(map[string]interface{})
-
-				listener.HealthCheck = &appmesh.HealthCheckPolicy{}
-
-				if vHealthyThreshold, ok := mHealthCheck["healthy_threshold"].(int); ok && vHealthyThreshold > 0 {
-					listener.HealthCheck.HealthyThreshold = aws.Int64(int64(vHealthyThreshold))
-				}
-				if vIntervalMillis, ok := mHealthCheck["interval_millis"].(int); ok && vIntervalMillis > 0 {
-					listener.HealthCheck.IntervalMillis = aws.Int64(int64(vIntervalMillis))
-				}
-				if vPath, ok := mHealthCheck["path"].(string); ok && vPath != "" {
-					listener.HealthCheck.Path = aws.String(vPath)
-				}
-				if vPort, ok := mHealthCheck["port"].(int); ok && vPort > 0 {
-					listener.HealthCheck.Port = aws.Int64(int64(vPort))
-				}
-				if vProtocol, ok := mHealthCheck["protocol"].(string); ok && vProtocol != "" {
-					listener.HealthCheck.Protocol = aws.String(vProtocol)
-				}
-				if vTimeoutMillis, ok := mHealthCheck["timeout_millis"].(int); ok && vTimeoutMillis > 0 {
-					listener.HealthCheck.TimeoutMillis = aws.Int64(int64(vTimeoutMillis))
-				}
-				if vUnhealthyThreshold, ok := mHealthCheck["unhealthy_threshold"].(int); ok && vUnhealthyThreshold > 0 {
-					listener.HealthCheck.UnhealthyThreshold = aws.Int64(int64(vUnhealthyThreshold))
-				}
-			}
-
-			if vPortMapping, ok := mListener["port_mapping"].([]interface{}); ok && len(vPortMapping) > 0 && vPortMapping[0] != nil {
-				mPortMapping := vPortMapping[0].(map[string]interface{})
-
-				listener.PortMapping = &appmesh.PortMapping{}
-
-				if vPort, ok := mPortMapping["port"].(int); ok && vPort > 0 {
-					listener.PortMapping.Port = aws.Int64(int64(vPort))
-				}
-				if vProtocol, ok := mPortMapping["protocol"].(string); ok && vProtocol != "" {
-					listener.PortMapping.Protocol = aws.String(vProtocol)
-				}
-			}
-
-			listeners = append(listeners, listener)
-		}
-
-		spec.Listeners = listeners
-	}
-
-	if vLogging, ok := mSpec["logging"].([]interface{}); ok && len(vLogging) > 0 && vLogging[0] != nil {
-		mLogging := vLogging[0].(map[string]interface{})
-
-		if vAccessLog, ok := mLogging["access_log"].([]interface{}); ok && len(vAccessLog) > 0 && vAccessLog[0] != nil {
-			mAccessLog := vAccessLog[0].(map[string]interface{})
-
-			if vFile, ok := mAccessLog["file"].([]interface{}); ok && len(vFile) > 0 && vFile[0] != nil {
-				mFile := vFile[0].(map[string]interface{})
-
-				if vPath, ok := mFile["path"].(string); ok && vPath != "" {
-					spec.Logging = &appmesh.Logging{
-						AccessLog: &appmesh.AccessLog{
-							File: &appmesh.FileAccessLog{
-								Path: aws.String(vPath),
-							},
-						},
-					}
-				}
-			}
-		}
-	}
-
-	if vServiceDiscovery, ok := mSpec["service_discovery"].([]interface{}); ok && len(vServiceDiscovery) > 0 && vServiceDiscovery[0] != nil {
-		spec.ServiceDiscovery = &appmesh.ServiceDiscovery{}
-
-		mServiceDiscovery := vServiceDiscovery[0].(map[string]interface{})
-
-		if vAwsCloudMap, ok := mServiceDiscovery["aws_cloud_map"].([]interface{}); ok && len(vAwsCloudMap) > 0 && vAwsCloudMap[0] != nil {
-			spec.ServiceDiscovery.AwsCloudMap = &appmesh.AwsCloudMapServiceDiscovery{}
-
-			mAwsCloudMap := vAwsCloudMap[0].(map[string]interface{})
-
-			if vAttributes, ok := mAwsCloudMap["attributes"].(map[string]interface{}); ok && len(vAttributes) > 0 {
-				attributes := []*appmesh.AwsCloudMapInstanceAttribute{}
-
-				for k, v := range vAttributes {
-					attributes = append(attributes, &appmesh.AwsCloudMapInstanceAttribute{
-						Key:   aws.String(k),
-						Value: aws.String(v.(string)),
-					})
-				}
-
-				spec.ServiceDiscovery.AwsCloudMap.Attributes = attributes
-			}
-			if vNamespaceName, ok := mAwsCloudMap["namespace_name"].(string); ok && vNamespaceName != "" {
-				spec.ServiceDiscovery.AwsCloudMap.NamespaceName = aws.String(vNamespaceName)
-			}
-			if vServiceName, ok := mAwsCloudMap["service_name"].(string); ok && vServiceName != "" {
-				spec.ServiceDiscovery.AwsCloudMap.ServiceName = aws.String(vServiceName)
-			}
-		}
-
-		if vDns, ok := mServiceDiscovery["dns"].([]interface{}); ok && len(vDns) > 0 && vDns[0] != nil {
-			mDns := vDns[0].(map[string]interface{})
-
-			if vHostname, ok := mDns["hostname"].(string); ok && vHostname != "" {
-				spec.ServiceDiscovery.Dns = &appmesh.DnsServiceDiscovery{
-					Hostname: aws.String(vHostname),
-				}
-			}
-		}
-	}
-
-	return spec
-}
-
-func flattenAppmeshVirtualNodeSpec(spec *appmesh.VirtualNodeSpec) []interface{} {
-	if spec == nil {
-		return []interface{}{}
-	}
-
-	mSpec := map[string]interface{}{}
-
-	if spec.Backends != nil {
-		vBackends := []interface{}{}
-
-		for _, backend := range spec.Backends {
-			mBackend := map[string]interface{}{}
-
-			if backend.VirtualService != nil {
-				mVirtualService := map[string]interface{}{
-					"virtual_service_name": aws.StringValue(backend.VirtualService.VirtualServiceName),
-				}
-				mBackend["virtual_service"] = []interface{}{mVirtualService}
-			}
-
-			vBackends = append(vBackends, mBackend)
-		}
-
-		mSpec["backend"] = schema.NewSet(appmeshVirtualNodeBackendHash, vBackends)
-	}
-
-	if spec.Listeners != nil {
-		vListeners := []interface{}{}
-
-		for _, listener := range spec.Listeners {
-			mListener := map[string]interface{}{}
-
-			if listener.HealthCheck != nil {
-				mHealthCheck := map[string]interface{}{
-					"healthy_threshold":   int(aws.Int64Value(listener.HealthCheck.HealthyThreshold)),
-					"interval_millis":     int(aws.Int64Value(listener.HealthCheck.IntervalMillis)),
-					"path":                aws.StringValue(listener.HealthCheck.Path),
-					"port":                int(aws.Int64Value(listener.HealthCheck.Port)),
-					"protocol":            aws.StringValue(listener.HealthCheck.Protocol),
-					"timeout_millis":      int(aws.Int64Value(listener.HealthCheck.TimeoutMillis)),
-					"unhealthy_threshold": int(aws.Int64Value(listener.HealthCheck.UnhealthyThreshold)),
-				}
-				mListener["health_check"] = []interface{}{mHealthCheck}
-			}
-
-			if listener.PortMapping != nil {
-				mPortMapping := map[string]interface{}{
-					"port":     int(aws.Int64Value(listener.PortMapping.Port)),
-					"protocol": aws.StringValue(listener.PortMapping.Protocol),
-				}
-				mListener["port_mapping"] = []interface{}{mPortMapping}
-			}
-
-			vListeners = append(vListeners, mListener)
-		}
-
-		mSpec["listener"] = schema.NewSet(appmeshVirtualNodeListenerHash, vListeners)
-	}
-
-	if spec.Logging != nil && spec.Logging.AccessLog != nil && spec.Logging.AccessLog.File != nil {
-		mSpec["logging"] = []interface{}{
-			map[string]interface{}{
-				"access_log": []interface{}{
-					map[string]interface{}{
-						"file": []interface{}{
-							map[string]interface{}{
-								"path": aws.StringValue(spec.Logging.AccessLog.File.Path),
-							},
-						},
-					},
-				},
-			},
-		}
-	}
-
-	if spec.ServiceDiscovery != nil {
-		mServiceDiscovery := map[string]interface{}{}
-
-		if spec.ServiceDiscovery.AwsCloudMap != nil {
-			vAttributes := map[string]interface{}{}
-
-			for _, attribute := range spec.ServiceDiscovery.AwsCloudMap.Attributes {
-				vAttributes[aws.StringValue(attribute.Key)] = aws.StringValue(attribute.Value)
-			}
-
-			mServiceDiscovery["aws_cloud_map"] = []interface{}{
-				map[string]interface{}{
-					"attributes":     vAttributes,
-					"namespace_name": aws.StringValue(spec.ServiceDiscovery.AwsCloudMap.NamespaceName),
-					"service_name":   aws.StringValue(spec.ServiceDiscovery.AwsCloudMap.ServiceName),
-				},
-			}
-		}
-
-		if spec.ServiceDiscovery.Dns != nil {
-			mServiceDiscovery["dns"] = []interface{}{
-				map[string]interface{}{
-					"hostname": aws.StringValue(spec.ServiceDiscovery.Dns.Hostname),
-				},
-			}
-		}
-
-		mSpec["service_discovery"] = []interface{}{mServiceDiscovery}
-	}
-
-	return []interface{}{mSpec}
-}
-
-func expandAppmeshVirtualServiceSpec(vSpec []interface{}) *appmesh.VirtualServiceSpec {
-	spec := &appmesh.VirtualServiceSpec{}
-
-	if len(vSpec) == 0 || vSpec[0] == nil {
-		// Empty Spec is allowed.
-		return spec
-	}
-	mSpec := vSpec[0].(map[string]interface{})
-
-	if vProvider, ok := mSpec["provider"].([]interface{}); ok && len(vProvider) > 0 && vProvider[0] != nil {
-		mProvider := vProvider[0].(map[string]interface{})
-
-		spec.Provider = &appmesh.VirtualServiceProvider{}
-
-		if vVirtualNode, ok := mProvider["virtual_node"].([]interface{}); ok && len(vVirtualNode) > 0 && vVirtualNode[0] != nil {
-			mVirtualNode := vVirtualNode[0].(map[string]interface{})
-
-			if vVirtualNodeName, ok := mVirtualNode["virtual_node_name"].(string); ok && vVirtualNodeName != "" {
-				spec.Provider.VirtualNode = &appmesh.VirtualNodeServiceProvider{
-					VirtualNodeName: aws.String(vVirtualNodeName),
-				}
-			}
-		}
-
-		if vVirtualRouter, ok := mProvider["virtual_router"].([]interface{}); ok && len(vVirtualRouter) > 0 && vVirtualRouter[0] != nil {
-			mVirtualRouter := vVirtualRouter[0].(map[string]interface{})
-
-			if vVirtualRouterName, ok := mVirtualRouter["virtual_router_name"].(string); ok && vVirtualRouterName != "" {
-				spec.Provider.VirtualRouter = &appmesh.VirtualRouterServiceProvider{
-					VirtualRouterName: aws.String(vVirtualRouterName),
-				}
-			}
-		}
-	}
-
-	return spec
-}
-
-func flattenAppmeshVirtualServiceSpec(spec *appmesh.VirtualServiceSpec) []interface{} {
-	if spec == nil {
-		return []interface{}{}
-	}
-
-	mSpec := map[string]interface{}{}
-
-	if spec.Provider != nil {
-		mProvider := map[string]interface{}{}
-
-		if spec.Provider.VirtualNode != nil {
-			mProvider["virtual_node"] = []interface{}{
-				map[string]interface{}{
-					"virtual_node_name": aws.StringValue(spec.Provider.VirtualNode.VirtualNodeName),
-				},
-			}
-		}
-
-		if spec.Provider.VirtualRouter != nil {
-			mProvider["virtual_router"] = []interface{}{
-				map[string]interface{}{
-					"virtual_router_name": aws.StringValue(spec.Provider.VirtualRouter.VirtualRouterName),
-				},
-			}
-		}
-
-		mSpec["provider"] = []interface{}{mProvider}
-	}
-
-	return []interface{}{mSpec}
 }
 
 func flattenAppsyncPipelineConfig(c *appsync.PipelineConfig) []interface{} {
